% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allapps3.R
\docType{data}
\name{LPA.app}
\alias{LPA.app}
\title{Launch a shiny-app simulating the LPA model}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
LPA.app
}
\description{
Launch a shiny-app simulating the LPA model
}
\details{
Launch app for details
}
\examples{
\dontrun{LPA.app}
}
\keyword{datasets}
