% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allapps3.R
\docType{data}
\name{LVcomp.app}
\alias{LVcomp.app}
\title{Launch a shiny-app simulating the Lotka-Volterra competition model}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
LVcomp.app
}
\description{
Launch a shiny-app simulating the Lotka-Volterra competition model
}
\details{
Launch app for details
}
\examples{
\dontrun{LVcomp.app}
}
\keyword{datasets}
