% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allapps3.R
\docType{data}
\name{LVpred.app}
\alias{LVpred.app}
\title{Launch a shiny-app simulating the Lotka-Volterra predation model}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
LVpred.app
}
\description{
Launch a shiny-app simulating the Lotka-Volterra predation model
}
\details{
Launch app for details
}
\examples{
\dontrun{LVpred.app}
}
\keyword{datasets}
