% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allapps3.R
\docType{data}
\name{May.app}
\alias{May.app}
\title{Launch a shiny-app simulating May's Parasitoid-host Model model}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
May.app
}
\description{
Launch a shiny-app simulating May's Parasitoid-host Model model
}
\details{
Launch app for details
}
\examples{
\dontrun{May.app}
}
\keyword{datasets}
