% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allapps3.R
\docType{data}
\name{rosenzweigmacarthur.app}
\alias{rosenzweigmacarthur.app}
\title{Launch a shiny-app simulating the Rosenzweig-MacArthur model}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
rosenzweigmacarthur.app
}
\description{
Launch a shiny-app simulating the Rosenzweig-MacArthur model
}
\details{
Launch app for details
}
\examples{
\dontrun{rosenzweigmacarthur.app}
}
\keyword{datasets}
