% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allapps3.R
\docType{data}
\name{SEIR.app}
\alias{SEIR.app}
\title{Launch a shiny-app simulating the seasonal SEIR model}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
SEIR.app
}
\description{
Launch a shiny-app simulating the seasonal SEIR model
}
\details{
Launch app for details
}
\examples{
\dontrun{SEIR.app}
}
\keyword{datasets}
