% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allapps3.R
\docType{data}
\name{SEIRS.app}
\alias{SEIRS.app}
\title{Launch a shiny-app simulating the SEIRS model}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
SEIRS.app
}
\description{
Launch a shiny-app simulating the SEIRS model
}
\details{
Launch app for details
}
\examples{
\dontrun{SEIRS.app}
}
\keyword{datasets}
