% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allapps3.R
\docType{data}
\name{SIR.app}
\alias{SIR.app}
\title{Launch a shiny-app simulating the SIR model}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
SIR.app
}
\description{
Launch a shiny-app simulating the SIR model
}
\details{
Launch app for details
}
\examples{
\dontrun{SIR.app}
}
\keyword{datasets}
