% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allapps3.R
\docType{data}
\name{TSIR.app}
\alias{TSIR.app}
\title{Launch a shiny-app simulating TSIR model}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
TSIR.app
}
\description{
Launch a shiny-app simulating TSIR model
}
\details{
Launch app for details
}
\examples{
\dontrun{TSIR.app}
}
\keyword{datasets}
